backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pymysql==1.1.1
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.35.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
httpx==0.27.0
aiosmtplib==3.0.1
email-validator==2.1.1
pyotp==2.9.0
cryptography==42.0.7
razorpay>=1.4.1  # optional: for Razorpay UPI gateway
pymongo>=4.6.1